figured out how to store cube positions
stored in allPositions list in World component (note that this is NOT a state variable because it should not trigger a component re-render)

also the position is ONLY updated at step "0_ROTATION" and is not updated in the middle of a rotation animation. Therefore, this solution should only work if you are moving one cube at a time. If multiple cubes are moving, positions may have to be updated more frequently than this