fixed the bug with axisOfRotationWorld vs Local
Previously the cube might rotate in the wrong direction. Basically the UI picks the axis based on the world axes, but the cube rotates around the axis based on its local axis. So you have to flip the axis of rotation in the cases where the local and world are opposite of each other (can happen through rotations as the cube changes its orientation around in 3D space).